revert bad merge merge clone-5.6.17 into 7.3.5

+++ storage/innobase/pars/pars0pars.cc
@@ -2160,9 +2160,8 @@ pars_get_lex_chars(
 {
 	int	len;
 
-	len = static_cast<int>(
-		pars_sym_tab_global->string_len
-		- pars_sym_tab_global->next_char_pos);
+	len = pars_sym_tab_global->string_len
+		- pars_sym_tab_global->next_char_pos;
 	if (len == 0) {
 #ifdef YYDEBUG
 		/* fputs("SQL string ends\n", stderr); */
