+++ storage/innobase/trx/trx0trx.cc
@@ -1809,8 +1809,7 @@ state_ok:
 	}
 
 	if (trx->mysql_thd != NULL) {
-		innobase_mysql_print_thd(
-			f, trx->mysql_thd, static_cast<uint>(max_query_len));
+		innobase_mysql_print_thd(f, trx->mysql_thd, max_query_len);
 	}
 }
